Clone node wiki pages when forking/registering node
Instead of copying over the list of node wiki page
versions to the fork or registration, clone each
node wiki page and add them to the fork or registration
wiki_pages_versions list. If the version is the current
wiki version, add it's cloned copy to the fork or
registration's wiki_pages_current field. Also set
wiki_private_uuids = {} for forks and registrations.